fixed bug for rating decimal. fixed js for save review. refactor ability for new and create. added user_decorator for set role of new user.

+++ db/migrate/20081020220724_create_reviews.rb
@@ -2,7 +2,7 @@ class CreateReviews < ActiveRecord::Migration
   def self.up
     create_table :reviews do |t|
       t.string  :name, :location
-      t.integer :rating
+      t.decimal :rating
       t.text :review
       t.boolean :approved, :default => false
